project/requirements3.12.7.txt
absl-py==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
antlr4-python3-runtime==4.9.3
anyio==4.13.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.3.0
attrs==26.1.0
babel==2.18.0
beautifulsoup4==4.14.3
biopython==1.87
bleach==6.3.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.7
cuda-bindings==13.2.0
cuda-pathfinder==1.5.3
cuda-toolkit==13.0.2
cycler==0.12.1
Cython==3.2.4
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.1
einops==0.8.2
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.29.0
fonttools==4.62.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2026.3.0
gitdb==4.0.12
GitPython==3.1.47
graphql-core==3.2.8
GridDataFormats==1.1.0
grpcio==1.80.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.11.0
idna==3.11
ipykernel==7.2.0
ipython==9.12.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
json5==0.14.0
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter-events==0.12.1
jupyter-lsp==2.3.1
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kiwisolver==1.5.0
lark==1.3.1
lightning-utilities==0.15.3
llvmlite==0.47.0
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.9
matplotlib-inline==0.2.1
mda-xdrlib==0.2.0
MDAnalysis==2.10.0
mdurl==0.1.2
mistune==3.2.0
mmtf-python==1.1.3
more-itertools==11.0.2
mpmath==1.3.0
mrcfile==1.5.4
msgpack==1.1.2
multidict==6.7.1
multiprocess==0.70.19
nbclient==0.10.4
nbconvert==7.17.1
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
nodeenv==1.10.0
notebook_shim==0.2.4
numba==0.65.1
numpy==2.4.4
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
omegaconf==2.3.0
opendock==0.0.1
orjson==3.11.8
packaging==26.1
pandas==3.0.2
pandocfilters==1.5.1
parso==0.8.6
pathos==0.3.5
pexpect==4.9.0
pillow==12.2.0
platformdirs==4.9.6
pox==0.3.7
ppft==1.7.8
ProDy==2.6.1
prometheus_client==0.25.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==7.34.1
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==3.0
Pygments==2.20.0
pynndescent==0.6.0
pyparsing==3.1.1
pyright==1.1.408
pyrosetta==2026.3+releasequarterly.5e498f1409
python-dateutil==2.9.0.post0
python-json-logger==4.1.0
pytorch-lightning==2.6.1
PyYAML==6.0.3
pyzmq==27.1.0
rcsb-api==1.7.1
rdkit==2026.3.1
referencing==0.37.0
requests==2.33.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==15.0.0
rpds-py==0.30.0
rustworkx==0.17.1
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
Send2Trash==2.1.0
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
soupsieve==2.8.3
stack-data==0.6.3
sympy==1.14.0
-e git+ssh://git@github.com/adsfibonacci/synformer.git@64d119474a70d5ef30265db9f037460f69c4bb1b#egg=synformer&subdirectory=../../../../project/dependencies/synformer
tensorboard==2.20.0
tensorboard-data-server==0.7.2
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
torch==2.11.0
torchmetrics==1.9.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
triton==3.6.0
typer==0.24.2
types-PyYAML==6.0.12.20260408
types-setuptools==82.0.0.20260408
typing_extensions==4.15.0
tzdata==2026.1
umap-learn==0.5.12
unidock_tools @ git+https://github.com/dptech-corp/Uni-Dock.git@6c4c72da22755b8dcb336c18b96ed62d01993c2f#subdirectory=unidock_tools
uri-template==1.3.0
urllib3==2.6.3
wcwidth==0.6.0
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.8
widgetsnbextension==4.0.15
yarl==1.23.0
